+++ Sources/Sentry/SentrySDKInternal.m
@@ -1,7 +1,6 @@
 #import "SentrySDKInternal.h"
 #import "PrivateSentrySDKOnly.h"
 #import "SentryAppStartMeasurement.h"
-#import "SentryAutoBreadcrumbTrackingIntegration.h"
 #import "SentryBreadcrumb.h"
 #import "SentryClient+Private.h"
 #import "SentryCoreDataTrackingIntegration.h"
@@ -500,8 +499,7 @@ static NSDate *_Nullable startTimestamp = nil;
 #if SENTRY_HAS_UIKIT
             [SentryAppStartTrackingIntegration class],
 #endif // SENTRY_HAS_UIKIT
-        [SentryAutoBreadcrumbTrackingIntegration class], [SentryCoreDataTrackingIntegration class],
-        nil];
+        [SentryCoreDataTrackingIntegration class], nil];
 
     return defaultIntegrations;
 }