+++ Package.swift
@@ -22,28 +22,28 @@ var products: [Product] = [
 var targets: [Target] = [
     .binaryTarget(
         name: "Sentry",
-        url: "https://github.com/getsentry/sentry-cocoa/releases/download/9.2.0/Sentry.xcframework.zip",
-        checksum: "3f68af784da3437aeb24527a7353d39bbcd2fe033ffbb37b7ab687b699e8982b" //Sentry-Static
+        url: "https://github.com/getsentry/sentry-cocoa/releases/download/9.3.0/Sentry.xcframework.zip",
+        checksum: "58098bef1eb0cdfcd3a070fa84097ece82e0b6a9c7a17148522cc2e07ddf45d7" //Sentry-Static
     ),
     .binaryTarget(
         name: "Sentry-Dynamic",
-        url: "https://github.com/getsentry/sentry-cocoa/releases/download/9.2.0/Sentry-Dynamic.xcframework.zip",
-        checksum: "370d1d5ffec17114697ac5a6ee931d4c39783c1c656456fe09c91e8838ab6fe2" //Sentry-Dynamic
+        url: "https://github.com/getsentry/sentry-cocoa/releases/download/9.3.0/Sentry-Dynamic.xcframework.zip",
+        checksum: "30292d379cf14bf232d51b1b700a40801e8185431b7bb04e2bfe787d07aeb599" //Sentry-Dynamic
     ),
     .binaryTarget(
         name: "Sentry-Dynamic-WithARM64e",
-        url: "https://github.com/getsentry/sentry-cocoa/releases/download/9.2.0/Sentry-Dynamic-WithARM64e.xcframework.zip",
-        checksum: "682a96f23954de5c723398b4d09ff1c70279bc58edc900db2214a7fb19fea75b" //Sentry-Dynamic-WithARM64e
+        url: "https://github.com/getsentry/sentry-cocoa/releases/download/9.3.0/Sentry-Dynamic-WithARM64e.xcframework.zip",
+        checksum: "270805b181e544b448ba245337c415b9b2952e4103b600dd9ce2c38d7176bec9" //Sentry-Dynamic-WithARM64e
     ),
     .binaryTarget(
         name: "Sentry-WithoutUIKitOrAppKit",
-        url: "https://github.com/getsentry/sentry-cocoa/releases/download/9.2.0/Sentry-WithoutUIKitOrAppKit.xcframework.zip",
-        checksum: "24d88c9ddb23bb9799ec4ccf7fa32de0b604e4735e6b2d87846191f2506a48f4" //Sentry-WithoutUIKitOrAppKit
+        url: "https://github.com/getsentry/sentry-cocoa/releases/download/9.3.0/Sentry-WithoutUIKitOrAppKit.xcframework.zip",
+        checksum: "37f61b11bdf19a8e88cfd9b7e9a500ec38f4073b405d20d7efce565b4d8aa1bd" //Sentry-WithoutUIKitOrAppKit
     ),
     .binaryTarget(
         name: "Sentry-WithoutUIKitOrAppKit-WithARM64e",
-        url: "https://github.com/getsentry/sentry-cocoa/releases/download/9.2.0/Sentry-WithoutUIKitOrAppKit-WithARM64e.xcframework.zip",
-        checksum: "8ad2e8a98e96cb613d16183c44f18ca32c96dbc0c5ec6999b21c020f7148310b" //Sentry-WithoutUIKitOrAppKit-WithARM64e
+        url: "https://github.com/getsentry/sentry-cocoa/releases/download/9.3.0/Sentry-WithoutUIKitOrAppKit-WithARM64e.xcframework.zip",
+        checksum: "05ce39a7243b70d3cc01f9f1e94cd823d3c1b5c93ce23525814cff5d3e9cff9b" //Sentry-WithoutUIKitOrAppKit-WithARM64e
     ),
     .target(
         name: "SentrySwiftUI",